Lots of stuff (partial support, category and article module, language routing)

+++ application/models/ArticleModel.php
@@ -12,7 +12,7 @@ class ArticleModel extends AbstractModel
             JOIN category_content cc ON (cc.category_id = a.category_id)
             WHERE ac.language_id = :language_id AND cc.language_id = :language_id
             ORDER BY a.publish_date
-        ', array('language_id' => array('value' => 1, 'type' => PDO::PARAM_INT)));
+        ', array('language_id' => array('value' => $_SESSION['language_id'], 'type' => PDO::PARAM_INT)));
     }
 
     public function getSingle($id)
@@ -25,20 +25,21 @@ class ArticleModel extends AbstractModel
         ', array('id' => array('value' => $id, 'type' => PDO::PARAM_INT)));
 
         $languages = DBAdapter::getInstance()->query('
-            SELECT ac.language_id, ac.title, ac.content
+            SELECT ac.language_id, ac.title, ac.content, ac.is_online
             FROM article_content ac
             WHERE ac.article_id = :id
         ', array('id' => array('value' => $id, 'type' => PDO::PARAM_INT)));
 
         foreach ($languages as $language) {
-            $result[0]->language[$language->language_id] = array('title' => $language->title, 'content' => $language->content);
+            $result[0]->language[$language->language_id] = array('title' => $language->title, 'content' => $language->content, 'is_online' => $language->is_online);
         }
 
         return $result[0];
     }
 
-    public function saveArticle($postdata, $id)
+    public function saveArticle($postdata, $id = null)
     {
+        //var_dump($postdata['language']); exit;
         $article = array();
         foreach ($postdata as $key => $value) {
             if ($key == 'save' || $key == 'language' || $key == 'languages') {
@@ -47,12 +48,18 @@ class ArticleModel extends AbstractModel
             $article[$key] = $value;
         }
 
-        DBAdapter::getInstance()->update('article', $article, array('id' => array('value' => $id, 'type' => PDO::PARAM_INT)));
-        DBAdapter::getInstance()->delete('article_content', array('article_id' => array('value' => $id, 'type' => PDO::PARAM_INT)));
-        foreach ($postdata['languages'] as $language_id) {
+        if ($id === null) {
+            $id = DBAdapter::getInstance()->insert('article', $article);
+        } else {
+            DBAdapter::getInstance()->update('article', $article, array('id' => array('value' => $id, 'type' => PDO::PARAM_INT)));
+            DBAdapter::getInstance()->delete('article_content', array('article_id' => array('value' => $id, 'type' => PDO::PARAM_INT)));
+        }
+
+        foreach ($postdata['language'] as $language_id => $language) {
             //Set foreign keys correctly
             $postdata['language'][$language_id]['language_id'] = $language_id;
             $postdata['language'][$language_id]['article_id'] = $id;
+
             DBAdapter::getInstance()->insert('article_content', $postdata['language'][$language_id]);
         }
 
@@ -67,7 +74,10 @@ class ArticleModel extends AbstractModel
             JOIN news_categories_content ncac ON (ncac.news_category_id = n.news_category_id)
             WHERE nco.language_id = :language_id AND ncac.language_id = :language_id
             ORDER BY n.publish_date DESC LIMIT :limit
-        ', array('language_id' => array('value' => 1, 'type' => PDO::PARAM_INT),  'limit' => array('value' => $this->model_config['limit'], 'type' => PDO::PARAM_INT)));
+        ', array(
+            'language_id'   => array('value' => 1, 'type' => PDO::PARAM_INT),
+            'limit'         => array('value' => $this->model_config['limit'], 'type' => PDO::PARAM_INT)
+        ));
     }
 
 }
